requirements.txt
pymongo==4.2.0
pytest==7.1.3
pyyaml==6.0
requests==2.28.2
black==22.8.0
numpy==1.24.1
rich==13.3.1
python-ulid==1.1.0
statbotics==3.0.0
pyfakefs==5.3.2
pandas==2.0.3
statsmodels==0.14.1
dnspython==2.6.0
rich==13.3.1
httpx==0.28.1